requirements.txt
amqp==1.4.8
anyjson==0.3.3
certifi==2018.4.16
chardet==3.0.4
Django==1.8.4
django-extensions==1.5.7
django-storages==1.1.8
djangorestframework==3.2.4
enum34==1.0.4
idna==2.7
MySQL-python==1.2.5
pytz==2015.7
requests==2.19.1
six==1.11.0
urllib3==1.23
